Add py3.3 to supported/tested lists
The C-module is written against the 3.3+ API. We could skip testing this
and just declare it's 3.4+ (it'd certainly be easier given all the hoops
one has to jump through to get tox working on 3.3) but I like to support
as many versions as I can where possible.

Still - do feel free to reject this commit - I can always strip it out
easily enough.