Parity work: don't call tell() in exception handlers
This is a modification of the Python version to make it more like the C
version:

We don't call tell() in the C version in the event of an exception and
arguably we shouldn't anyway as there's no guarantee that all file-like
objects implement (e.g. sockets and pipes). This might make debugging a
little harder for users but they can always call it themselves if they
need it, and this makes the decoder applicable to literally any readable
file-like object.