Parity work: make _find_encoders equivalent
Previously, the C version was actively importing deferred types. The
Python algorithm was better, only checking if the specified module was
*already* imported (which, if we're encoding the specified type it must
be or the specified type couldn't exist yet).

This commit also removes the ugly reference counting in the deferred
logic, using the standard "borrowed references on entry, new references
on return" idiom throughout instead.