Parity work: add str_errors, tweak indefinite str decoding
The C variant has a "str_errors" parameter, permitting configuration of
the error-handling mode used when decoding UTF-8 strings. This commit
adds this attribute to the Python variant, and also updates the decoding
of indefinite-length byte and unicode strings to match the algorithm in
the C variant.

Specifically, this disallows split-unicode characters in indefinite
length unicode strings, and also rejects non-string elements in
indefinite length byte and unicode strings.