Fix final windows build bugs
Okay, I confess I'm utterly stumped on this one. Took ages to find as
despite my best efforts I've no idea how to interactively debug stuff
under VS (after installing the lib via pip), so this was back to prints
and a certain amount of guess-work. Anyway...

Turns out the "s#" format (for string+length) in Py_BuildValue (or its
derivatives like PyObject_CallMethod, etc.) doesn't work on Windows. It
just results in a MemoryError whenever it's used! However, the simple
"s" format (for NULL-terminated string) works fine. It's not exactly a
huge deal to have it calculate the length of "big" so this is a
worthwhile trade-off to get stuff working on Windows. At this point all
the tests pass in both the C and Python variants.