Set up CI with Azure Pipelines
This is a squash of my earlier mess of commits; this configures Azure
Pipelines to test cbor2 under Linux, Mac OS X, and a couple of versions
of Visual Studio under Windows. It tests versions 2.7, 3.4 (on Linux),
3.5, 3.6, 3.7, pypy (2.x) and pypy3.

Coverage analysis is left to an external tool, but test results are
published to pipelines.