Add separate timezone sign handling
Previously the sign was bundled in with the hour. The issue with this is that if the hour is zero, then the sign gets lost.

This adds the sign as a separate group to the regex, and then applies it to either the hours or minutes, depending on if hours is zero.