Ensure hours and minutes of timezone are affected by the sign
Previously, the sign and hour were linked, which meant that if the hour was zero, the sign was lost. This was fixed in https://github.com/agronholm/cbor2/issues/163

But, the minute should also be affected by the sign, since `timedelta(hours=-1, minutes=1) is `-0:59`, not `-1:01`.

This PR proposes a fix for this issue.